Create StudentSignup module
