+++ spec/handlers/newflow/verify_email_by_code_spec.rb
@@ -2,41 +2,56 @@ require 'rails_helper'
 
 module Newflow
   describe VerifyEmailByCode, type: :handler do
-    context 'when success' do
-      let(:params) do
-        { code: email.confirmation_code }
+    subject(:handler_call) { described_class.call(params: params) }
+
+    let(:params) { { code: email.confirmation_code } }
+    let(:email) { FactoryBot.create(:email_address, user: user) }
+    let(:user) { FactoryBot.create(:user, state: User::UNVERIFIED, role: role) }
+
+      context 'when student' do
+        let(:role) { User.roles[:student] }
+
+        it 'verifies the email address found by the given code' do
+          expect(email.verified).to be(false)
+          handler_call
+          email.reload
+          expect(email.verified).to be(true)
+        end
+
+        it 'activates the user' do
+          expect_any_instance_of(StudentSignup::ActivateStudent).to receive(:exec).and_call_original
+          handler_call
+          user.reload
+          expect(user.state).to eq('activated')
+        end
+
+        it 'outputs the user' do
+          result = handler_call
+          expect(result.outputs.user).to eq(user)
+        end
       end
 
-      let(:email) do
-        FactoryBot.create(:email_address, user: user)
-      end
+      context 'when instructor' do
+        let(:role) { User.roles[:instructor] }
 
-      let(:user) do
-        FactoryBot.create(:user, state: 'unverified')
-      end
+        it 'verifies the email address found by the given code' do
+          allow_any_instance_of(EducatorSignup::ActivateAccount).to receive(:exec).with(user: user)
+          expect(email.verified).to be(false)
+          handler_call
+          email.reload
+          expect(email.verified).to be(true)
+        end
 
-      it 'verifies the email address found by the given code' do
-        expect(email.verified).to be(false)
-        described_class.call(params: params)
-        email.reload
-        expect(email.verified).to be(true)
-      end
+        it 'calls EducatorSignup::ActivateAccount' do
+          expect_any_instance_of(EducatorSignup::ActivateAccount).to receive(:exec).with(user: user)
+          handler_call
+        end
 
-      it 'activates the user' do
-        expect_any_instance_of(ActivateUser).to receive(:call).and_call_original
-        described_class.call(params: params)
-        user.reload
-        expect(user.state).to eq('activated')
+        it 'outputs the user' do
+          allow_any_instance_of(EducatorSignup::ActivateAccount).to receive(:exec).with(user: user)
+          outputs = handler_call.outputs
+          expect(outputs.user).to eq(user)
+        end
       end
-
-      it 'outputs the user' do
-        result = described_class.call(params: params)
-        expect(result.outputs.user).to eq(user)
-      end
-    end
-
-    context 'when failure' do
-      # TODO
     end
   end
-end
